Majorly refactored resource loading! removed handle system - too verbose and doesnt solve concrete problem
now instead you directly load the resrouces form the assets
free to choose how to save and organize
cleaned up logic by using cute headers
added asprite import
improved animation class
removed unneeded code due to refactor
removed standalone json - moved to fileimports

+++ src/gamedev/PathFinding.cpp
@@ -103,7 +103,7 @@ namespace magique
             if (dist < MAGIQUE_PATHFINDING_CELL_SIZE)
             {
                 // We are in the last cell
-                if (i == path.size() - 1)
+                if (i == static_cast<int>(path.size()) - 1)
                 {
                     return curr;
                 }
